:truck: chore(api) move tube-repository to `src/shared`
