+++ api/scripts/prod/target-profile-migrations/common.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
-import * as tubeRepository from '../../../lib/infrastructure/repositories/tube-repository.js';
 import * as skillRepository from '../../../src/shared/infrastructure/repositories/skill-repository.js';
+import * as tubeRepository from '../../../src/shared/infrastructure/repositories/tube-repository.js';
 
 let allSkills;
 let allTubes;
