+++ api/scripts/prod/target-profile-migrations/parse-xls-files-for-target-profiles-migration.js
@@ -16,8 +16,8 @@ import XLSX from 'xlsx';
 
 import { disconnect, knex } from '../../../db/knex-database-connection.js';
 import { normalizeAndRemoveAccents } from '../../../lib/domain/services/validation-treatments.js';
-import * as tubeRepository from '../../../lib/infrastructure/repositories/tube-repository.js';
 import { learningContentCache as cache } from '../../../src/shared/infrastructure/caches/learning-content-cache.js';
+import * as tubeRepository from '../../../src/shared/infrastructure/repositories/tube-repository.js';
 import { logger } from '../../../src/shared/infrastructure/utils/logger.js';
 import { autoMigrateTargetProfile } from './common.js';
