+++ api/src/certification/results/infrastructure/repositories/certified-profile-repository.js
@@ -5,10 +5,6 @@ import { LOCALE } from '../../../../shared/domain/constants.js';
 const { FRENCH_SPOKEN } = LOCALE;
 import { knex } from '../../../../../db/knex-database-connection.js';
 import * as knowledgeElementRepository from '../../../../../lib/infrastructure/repositories/knowledge-element-repository.js';
-import * as tubeRepository from '../../../../../lib/infrastructure/repositories/tube-repository.js';
-import * as areaRepository from '../../../../../src/shared/infrastructure/repositories/area-repository.js';
-import * as competenceRepository from '../../../../../src/shared/infrastructure/repositories/competence-repository.js';
-import * as skillRepository from '../../../../../src/shared/infrastructure/repositories/skill-repository.js';
 import { NotFoundError } from '../../../../shared/domain/errors.js';
 import {
   CertifiedArea,
@@ -17,6 +13,10 @@ import {
   CertifiedSkill,
   CertifiedTube,
 } from '../../../../shared/domain/read-models/CertifiedProfile.js';
+import * as areaRepository from '../../../../shared/infrastructure/repositories/area-repository.js';
+import * as competenceRepository from '../../../../shared/infrastructure/repositories/competence-repository.js';
+import * as skillRepository from '../../../../shared/infrastructure/repositories/skill-repository.js';
+import * as tubeRepository from '../../../../shared/infrastructure/repositories/tube-repository.js';
 
 const get = async function (certificationCourseId) {
   const certificationDatas = await knex
