+++ api/src/devcomp/infrastructure/repositories/training-trigger-repository.js
@@ -1,11 +1,11 @@
 import _ from 'lodash';
 
 import * as thematicRepository from '../../../../lib/infrastructure/repositories/thematic-repository.js';
-import * as tubeRepository from '../../../../lib/infrastructure/repositories/tube-repository.js';
 import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 import { NotFoundError } from '../../../shared/domain/errors.js';
 import * as areaRepository from '../../../shared/infrastructure/repositories/area-repository.js';
 import * as competenceRepository from '../../../shared/infrastructure/repositories/competence-repository.js';
+import * as tubeRepository from '../../../shared/infrastructure/repositories/tube-repository.js';
 import { logger } from '../../../shared/infrastructure/utils/logger.js';
 import { TrainingTrigger } from '../../domain/models/TrainingTrigger.js';
 import { TrainingTriggerTube } from '../../domain/models/TrainingTriggerTube.js';
