+++ api/src/prescription/target-profile/infrastructure/repositories/target-profile-administration-repository.js
@@ -2,7 +2,6 @@ import _ from 'lodash';
 
 import { knex } from '../../../../../db/knex-database-connection.js';
 import * as thematicRepository from '../../../../../lib/infrastructure/repositories/thematic-repository.js';
-import * as tubeRepository from '../../../../../lib/infrastructure/repositories/tube-repository.js';
 import { TargetProfileForAdmin } from '../../../../prescription/target-profile/domain/models/TargetProfileForAdmin.js';
 import { LOCALE } from '../../../../shared/domain/constants.js';
 import { constants } from '../../../../shared/domain/constants.js';
@@ -14,6 +13,7 @@ import { StageCollection } from '../../../../shared/domain/models/target-profile
 import * as areaRepository from '../../../../shared/infrastructure/repositories/area-repository.js';
 import * as competenceRepository from '../../../../shared/infrastructure/repositories/competence-repository.js';
 import * as skillRepository from '../../../../shared/infrastructure/repositories/skill-repository.js';
+import * as tubeRepository from '../../../../shared/infrastructure/repositories/tube-repository.js';
 import { TargetProfileSummaryForAdmin } from '../../domain/models/TargetProfileSummaryForAdmin.js';
 
 const { FRENCH_FRANCE } = LOCALE;
