+++ api/src/shared/infrastructure/repositories/tube-repository.js
@@ -1,8 +1,8 @@
-import { knex } from '../../../db/knex-database-connection.js';
-import { LearningContentResourceNotFound } from '../../../src/shared/domain/errors.js';
-import { Tube } from '../../../src/shared/domain/models/Tube.js';
-import { getTranslatedKey } from '../../../src/shared/domain/services/get-translated-text.js';
-import { LearningContentRepository } from '../../../src/shared/infrastructure/repositories/learning-content-repository.js';
+import { knex } from '../../../../db/knex-database-connection.js';
+import { LearningContentResourceNotFound } from '../../domain/errors.js';
+import { Tube } from '../../domain/models/Tube.js';
+import { getTranslatedKey } from '../../domain/services/get-translated-text.js';
+import { LearningContentRepository } from './learning-content-repository.js';
 
 const TABLE_NAME = 'learningcontent.tubes';
 const ACTIVE_STATUS = 'actif';
