+++ api/tests/shared/integration/infrastructure/repositories/tube-repository_test.js
@@ -1,6 +1,6 @@
-import * as tubeRepository from '../../../../lib/infrastructure/repositories/tube-repository.js';
-import { LearningContentResourceNotFound } from '../../../../src/shared/domain/errors.js';
-import { catchErr, databaseBuilder, domainBuilder, expect } from '../../../test-helper.js';
+import { LearningContentResourceNotFound } from '../../../../../src/shared/domain/errors.js';
+import * as tubeRepository from '../../../../../src/shared/infrastructure/repositories/tube-repository.js';
+import { catchErr, databaseBuilder, domainBuilder, expect } from '../../../../test-helper.js';
 
 describe('Integration | Repository | tube-repository', function () {
   const tubeData0 = {
